Add files via upload
This is an attempt to make a responsive website (different layout for different screen sizes). The website when opened on smartphones appears differently than on desktops. I was able to do this project after completing 10% of 'Web Development Bootcamp' on Udemy curated by Angela Yu. So far I am loving this.